Fix detection of aspect ratio.

+++ prototype1/src/defaultreconstructor.cpp
@@ -28,7 +28,7 @@ void DefaultReconstructor::processLine(const Line &line, PointCloud &cloud)
 	double alpha = M_PI_2 - dconf.projectorAngle;
 
 	for (const Line::Sample& sample : line.getSamples()) {
-		double rhp = (2.0 * ((double) sample.pos[0] / line.getResolution()[0]) - 1.0 + (1.0 / (2*line.getResolution()[0])))*(line.getResolution()[0]/line.getResolution()[1]);
+		double rhp = (2.0 * ((double) sample.pos[0] / line.getResolution()[0]) - 1.0 + (1.0 / (2*line.getResolution()[0])))*((double) line.getResolution()[0]/line.getResolution()[1]);
 		double rvp = -(2.0 * ((double) sample.pos[1] / line.getResolution()[1]) - 1.0 + (1.0 / (2*line.getResolution()[1])));
 		double fw = 1.0/std::tan(dconf.fov / 2.0);
 		double beta = M_PI_2 + std::atan(rvp / fw);
